CRUD  league + validaciones

+++ controllers/team.controller.js
@@ -1,6 +1,7 @@
 'use strict'
 var Team = require('../models/team.model');
 var League = require('../models/league.model');
+var User = require('../models/user.model');
 var jwt = require('../services/jwt');
 
 //SAVE TEAM 
@@ -20,35 +21,43 @@ function deleteTeam(req, res){
             if(err){
                res.status(500).send({message: 'error general', err})
             }else if(leagueFind){
-               if(leagueFind.admin == userId){
-           League.findOneAndUpdate({_id: leagueId, team: teamId},
-               {$pull: {team: teamId}}, {new:true}, (err, teamPull)=>{
-                   if(err){
-                       return res.status(500).send({message: 'Error general'})
-                   }else if(teamPull){
-                    Team.findByIdAndRemove(teamId, (err, teamRemoved)=>{
-                           if(err){
-                               return res.status(500).send({message: 'Error general', err})
-                           }else if(teamRemoved){
-                               return res.send({message: 'Equipo eliminado exitosamente', teamPull});
-                           }else{
-                               return res.status(404).send({message: 'Registro no encontrado o Equipo ya eliminado'})
-                           }
-                       })
-                   }else{
-                       return res.status(404).send({message: 'Eliminado de la base de datos, aun existente en Liga'})
-                   }
-               }).populate('team')
-       
-               }else{
-                   res.status(418).send({message: 'Usuario no permitido'});   
-               }
-            }else{
-                res.status(404).send({message: 'Liga no encontrada'});
-            }
-        })
+                User.findById(userId, (err, userFind)=>{
+                    if (err) {
+                        res.status(500).send({message: 'ERROR GENERAL', err})
+                    }else if (userFind) {
+                        if(leagueFind.admin == userId || userFind.role == 'ROLE_ADMIN'){
+                            League.findOneAndUpdate({_id: leagueId, team: teamId},
+                                {$pull: {team: teamId}}, {new:true}, (err, teamPull)=>{
+                                    if(err){
+                                        return res.status(500).send({message: 'Error general'})
+                                    }else if(teamPull){
+                                     Team.findByIdAndRemove(teamId, (err, teamRemoved)=>{
+                                            if(err){
+                                                return res.status(500).send({message: 'Error general', err})
+                                            }else if(teamRemoved){
+                                                return res.send({message: 'Equipo eliminado exitosamente', teamPull});
+                                            }else{
+                                                return res.status(404).send({message: 'Registro no encontrado o Equipo ya eliminado'})
+                                            }
+                                        })
+                                    }else{
+                                        return res.status(404).send({message: 'Eliminado de la base de datos, aun existente en Liga'})
+                                    }
+                                }).populate('team')
+                        
+                                }else{
+                                    res.status(418).send({message: 'Usuario no permitido'});   
+                                }
+                            }else{
+                                res.status(401).send({message: 'Usuario no econtrado'})
+                            }
+                        })
+                    }else{
+                        res.status(404).send({message: 'Liga no encontrada'});
+                    }
+                })
+        }
     }
-}
 //LIST TEAM
 function getTeams(req, res){
     Team.find({}).exec((err, team)=>{
